Refactor: update
