+++ core/src/com/mrwalker/firstgame/PlayerController/PlayerController.java
@@ -1,6 +1,6 @@
 package com.mrwalker.firstgame.PlayerController;
 
-import com.mrwalker.firstgame.Entity.Entity;
+import com.mrwalker.firstgame.Entity.Parts.Entity;
 
 public interface PlayerController {
     public void setController(Entity player);
